Start of Day commit
I have added a resource folder that holds a temporary .ico file to act as the icon for the installer. The base work for the scene transition has been set up, Originally I started making a new state for the event but I realized that I could just add a new update to our existing cinematic state.