fixes jobscheduler issues
Jobscheduler initialize was not acting as expected. This was due to the static keyword
being infront of most of the inline global variables and some functions. This is an
issue with c++17 or the author's (my) understanding of the keyword static. consider the
following:

	// foo.h
	static inline int foo;

	static inline Initialize()
	{
		foo = 56;
	}
	// main.cpp
	int main()
	{
		Initialize();
		foo = undefined;
	}
should foo be undefined here? I'm not certain but on Visual Studio 2019 with c++17 it is, so for
now I have opted to remove the static keywords and perform manual deallocation of all variables
in JobScheduler.h so they do not show up on the memory leak detection. (our memory leak detection
makes use of the fact that static variables must be destructed in the order they are constructed.
thus making variables static as well as inline combined with defining a static MemoryDumpOnDestruct
before all other variables allows us to hide static memory leak false positives)

-removes static keywords from most of the jobscheduler