End of week complete
Originally the orbit system would not represent the correct progression based on the level the player was on. (EX: if the player is goes from the tutorial level to level 2 there should be one planet in orbit but what we had would be there would be no planets in orbit) This set up in the previous example isn't game breaking but it could allow for the following example. If the player had finished the game, meaning they collected 3 planets, and then went to level 1. The planet count would not reset meaning that the player would be stuck on level 1 unless they quit the game or navigated using the level select. The system I set up today will be used to avoid this problem from arising.